Moved from react-navigation to expo-router, Moved RecipeCard back to components
