+++ .idea/workspace.xml
@@ -5,12 +5,17 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="b9b78b60-57b8-4638-9484-506b1ba2b05c" name="Changes" comment="">
-      <change afterPath="$PROJECT_DIR$/mobile-app/styles/EditRecipeFormStyles.ts" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/mobile-app/app/EditRecipeForm.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/mobile-app/app/EditRecipeForm.tsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/mobile-app/app/AddRecipeForm.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/mobile-app/app/EditRecipeForm.tsx" beforeDir="false" />
       <change beforePath="$PROJECT_DIR$/mobile-app/app/Home.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/mobile-app/app/Home.tsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/mobile-app/app/RecipeCard.tsx" beforeDir="false" />
       <change beforePath="$PROJECT_DIR$/mobile-app/app/RecipeDetail.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/mobile-app/app/RecipeDetail.tsx" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/mobile-app/app/_layout.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/mobile-app/app/_layout.tsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/mobile-app/app/ViewRecipes.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/mobile-app/app/ViewRecipes.tsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/mobile-app/app/navigation.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/mobile-app/app/navigation.tsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/mobile-app/package-lock.json" beforeDir="false" afterPath="$PROJECT_DIR$/mobile-app/package-lock.json" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/mobile-app/package.json" beforeDir="false" afterPath="$PROJECT_DIR$/mobile-app/package.json" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/mobile-app/styles/EditRecipeFormStyles.ts" beforeDir="false" afterPath="$PROJECT_DIR$/mobile-app/styles/EditRecipeFormStyles.ts" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -40,7 +45,7 @@
     &quot;cidr.known.project.marker&quot;: &quot;true&quot;,
     &quot;git-widget-placeholder&quot;: &quot;mobile-app-setup&quot;,
     &quot;kotlin-language-version-configured&quot;: &quot;true&quot;,
-    &quot;last_opened_file_path&quot;: &quot;C:/Users/o_ken/Desktop/GitHub Projects/RecipeStore&quot;,
+    &quot;last_opened_file_path&quot;: &quot;C:/Users/o_ken/Desktop/GitHub Projects/StickerSmash&quot;,
     &quot;project.structure.last.edited&quot;: &quot;Modules&quot;,
     &quot;project.structure.proportion&quot;: &quot;0.0&quot;,
     &quot;project.structure.side.proportion&quot;: &quot;0.0&quot;
@@ -50,6 +55,11 @@
     <key name="MoveFile.RECENT_KEYS">
       <recent name="C:\Users\o_ken\Desktop\GitHub Projects\RecipeStore\mobile-app\app" />
     </key>
+  </component>
+  <component name="RunAnythingCache">
+    <option name="myCommands">
+      <command value="y" />
+    </option>
   </component>
   <component name="SpellCheckerSettings" RuntimeDictionaries="0" Folders="0" CustomDictionaries="0" DefaultDictionary="application-level" UseSingleDictionary="true" transferred="true" />
   <component name="TaskManager">
